# api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Isidor API",
    "description": "\n    # Isidor Health Optimization API\n    \n    Isidor is an AI-driven life protocol system designed to optimize fitness, nutrition, sleep, cognitive performance, and overall well-being.\n    \n    ## Key Features\n    \n    * **Health Data Integration**: Track and analyze health metrics from various sources\n    * **Protocol System**: Follow structured frameworks for health optimization\n    * **AI-Powered Insights**: Get personalized health insights using Gemini AI\n    * **Vector Embeddings**: Utilize semantic search for finding patterns in health data\n    * **Privacy-First Design**: Secure storage of health data with encryption\n    \n    ## Authentication\n    \n    This API uses JWT-based authentication. To access protected endpoints:\n    \n    1. Obtain an access token using the `/api/v1/auth/login` endpoint\n    2. Include the token in the Authorization header: `Authorization: Bearer {token}`\n    3. Use the refresh token to obtain a new access token when it expires\n    \n    ## Rate Limiting\n    \n    API requests are rate-limited to prevent abuse. The current limit is 100 requests per hour per IP address.\n    \n    ## Data Privacy\n    \n    All sensitive health data is encrypted at rest using PostgreSQL's pgcrypto extension.\n    ",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/ai/protocols/{user_id}": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Get Protocol Recommendation",
        "description": "Generate a protocol recommendation using Gemini AI.",
        "operationId": "get_protocol_recommendation_api_v1_ai_protocols__user_id__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolRecommendationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Protocol Recommendation Api V1 Ai Protocols  User Id  Post"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create User Endpoint",
        "description": "Create new user.\n\nThis endpoint is public to allow user registration.",
        "operationId": "create_user_endpoint_api_v1_users__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Read User",
        "description": "Get user by ID.",
        "operationId": "read_user_api_v1_users__user_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update User Endpoint",
        "description": "Update a user.",
        "operationId": "update_user_endpoint_api_v1_users__user_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete User Endpoint",
        "description": "Delete a user.",
        "operationId": "delete_user_endpoint_api_v1_users__user_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "title": "Response Delete User Endpoint Api V1 Users  User Id  Delete"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/protocols/": {
      "get": {
        "tags": [
          "Protocols"
        ],
        "summary": "Read Protocols",
        "description": "Retrieve protocols.",
        "operationId": "read_protocols_api_v1_protocols__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "target_metric",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Metric"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Protocol"
                  },
                  "title": "Response Read Protocols Api V1 Protocols  Get"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Protocols"
        ],
        "summary": "Create Protocol Endpoint",
        "description": "Create new protocol.",
        "operationId": "create_protocol_endpoint_api_v1_protocols__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Protocol"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/protocols/{protocol_id}": {
      "get": {
        "tags": [
          "Protocols"
        ],
        "summary": "Read Protocol",
        "description": "Get protocol by ID.",
        "operationId": "read_protocol_api_v1_protocols__protocol_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "protocol_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Protocol Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Protocol"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Protocols"
        ],
        "summary": "Update Protocol Endpoint",
        "description": "Update a protocol.",
        "operationId": "update_protocol_endpoint_api_v1_protocols__protocol_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "protocol_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Protocol Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProtocolUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Protocol"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Protocols"
        ],
        "summary": "Delete Protocol Endpoint",
        "description": "Delete a protocol.",
        "operationId": "delete_protocol_endpoint_api_v1_protocols__protocol_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "protocol_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Protocol Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Delete Protocol Endpoint Api V1 Protocols  Protocol Id  Delete"
                }
              }
            }
          },
          "404": {
            "description": "Protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/user-protocols/": {
      "get": {
        "tags": [
          "User Protocols"
        ],
        "summary": "Read User Protocols",
        "description": "Retrieve user protocols.",
        "operationId": "read_user_protocols_api_v1_user_protocols__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProtocolWithProtocol"
                  },
                  "title": "Response Read User Protocols Api V1 User Protocols  Get"
                }
              }
            }
          },
          "404": {
            "description": "User protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user-protocols/active": {
      "get": {
        "tags": [
          "User Protocols"
        ],
        "summary": "Read Active User Protocols",
        "description": "Retrieve active user protocols.",
        "operationId": "read_active_user_protocols_api_v1_user_protocols_active_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserProtocolWithProtocol"
                  },
                  "type": "array",
                  "title": "Response Read Active User Protocols Api V1 User Protocols Active Get"
                }
              }
            }
          },
          "404": {
            "description": "User protocol not found"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/user-protocols/enroll": {
      "post": {
        "tags": [
          "User Protocols"
        ],
        "summary": "Enroll In Protocol",
        "description": "Enroll user in a protocol.",
        "operationId": "enroll_in_protocol_api_v1_user_protocols_enroll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProtocolCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProtocol"
                }
              }
            }
          },
          "404": {
            "description": "User protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/user-protocols/{user_protocol_id}/status": {
      "put": {
        "tags": [
          "User Protocols"
        ],
        "summary": "Update Status",
        "description": "Update the status of a user protocol.",
        "operationId": "update_status_api_v1_user_protocols__user_protocol_id__status_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_protocol_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Protocol Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProtocolStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProtocol"
                }
              }
            }
          },
          "404": {
            "description": "User protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user-protocols/{user_protocol_id}/progress": {
      "get": {
        "tags": [
          "User Protocols"
        ],
        "summary": "Get Progress",
        "description": "Get progress information for a user protocol.",
        "operationId": "get_progress_api_v1_user_protocols__user_protocol_id__progress_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_protocol_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Protocol Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProtocolProgress"
                }
              }
            }
          },
          "404": {
            "description": "User protocol not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProtocolRecommendationRequest": {
        "properties": {
          "health_goal": {
            "type": "string",
            "title": "Health Goal"
          },
          "current_metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Current Metrics"
          }
        },
        "type": "object",
        "required": [
          "health_goal",
          "current_metrics"
        ],
        "title": "ProtocolRecommendationRequest"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {

            },
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}